move rendering into a separate file
