+++ code/src/mandelview/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(MANDELVIEW_FILES
     "${CMAKE_CURRENT_SOURCE_DIR}/cpp/main.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/renderer.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/mandelview/renderer.hpp"
 )
 
 # add directory structure to IDEs
@@ -12,6 +14,7 @@ add_executable(mandelview ${MANDELVIEW_FILES} "${DEPS_DIR}/glad/src/glad.c")
 target_include_directories(mandelview
     PRIVATE
     "${DEPS_DIR}/glad/include"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/private"
 )
 
 target_link_libraries(mandelview
